requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==19.3.0
black==19.10b0
click==7.1.2
dj-database-url==0.5.0
Django==2.2.27
django-crispy-forms==1.7.0
django-heroku==0.3.1
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pathspec==0.8.0
psycopg2-binary
pylint==2.5.3
python-dotenv==0.14.0
pytz==2020.1
regex==2020.7.14
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.1.0
wrapt==1.12.1
